requirements.txt
aiogram>=3.0,<4.0
python-dotenv>=1.0
openai>=1.0
PyPDF2>=3.0
python-docx>=1.0
